server/requirements.txt
alembic==1.13.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
Deprecated==1.2.14
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.6.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==22.0.0
idna==3.7
importlib_resources==6.4.0
itsdangerous==2.1.2
Jinja2==3.1.4
jwcrypto==1.5.6
limits==3.12.0
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.0
pillow==10.3.0
pip-autoremove==0.10.0
pipdeptree==2.20.0
pycparser==2.21
Pygments==2.18.0
PyJWT==2.8.0
python-dotenv==1.0.1
redis==5.0.3
requests==2.31.0
rich==13.7.1
setuptools==69.5.1
simplejson==3.19.2
soupsieve==2.5
SQLAlchemy==2.0.29
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.3
wrapt==1.16.0
